+++ include/tools/types/fixed.hpp
@@ -14,7 +14,7 @@
     #include <tools/types/fixed/fixed-cpp98.hpp>
 #endif
 
-#if defined(dHAS_TYPE_TRAITS ) && defined(dHAS_RVALUE_REFERENCES)
+#ifdef dHAS_TYPE_TRAITS
     #include <type_traits>
 #else
     #include <tools/types/traits.hpp>