+++ docs/history.md
@@ -12,11 +12,20 @@
 ----------------------------------
 [2021y-04m-03d][23:09:37] #41_stage
 
-1) type_of_enum                        100
+1) обновлен: version                    119
+2) обновлен: types                      119
+
+3) добавлен tools/type_traits.hpp       012 PRE
+
+4) type_of_enum                         100
      - переехал из traits в common
 
-2) улучшение common                    107 PRE
-     - фикс бага: add_const_data       101
+5) улучшение common                     107 PRE
+     - фикс бага: add_const_data        101
+
+6) исправление features:                118 PRE
+     - фикс бага: dHAS_TYPE_TRAITS
+     - фикс бага: dTRAIT
 
 
 **Версия 1.1.8 (предварительная)**